Search and user integration (#19)
* init commit

* working on recommendations

* recommendations working

* ui work

* able to scroll through todos

* Ignore Lint Error

* Todo scrolls to selected index

* Add Homes carousel

* ui work

* add UserService.ts

* ui work

* add refresh functionality to todos and tenants as leases

* add search on tenants name to lease

* Todo card ui work

* fixed lint errors

* Variable naming changes

* Remove unnecessary observers

* made most recommended changes

* fixing more recommendations

* fixing issues

* finished fixing errors

* started analytics

* adding and getting transactions

* removed all unnecessary observers

* working on add transactions

* clean up

* finish transactions

* fix lint errors

---------

+++ src/Screens/AllProperties.tsx
@@ -8,16 +8,14 @@ import {SwipeListView} from "react-native-swipe-list-view";
 import { View, StyleSheet} from "react-native";
 import BackButton from "../Components/BackButton";
 import TrashButton from "../Components/TrashButton";
+import _ from "lodash";
 
 function AllProperties() {
 	const appContext = useAppContext();
 
-	const handleDeleteProperty = useCallback(async (propertyId:number) => {
-		try{
-			await appContext.deleteProperty(propertyId);
-		}catch (e){
-			return;
-		}
+	const handleDeleteProperty = useCallback(async (propertyId?:number) => {
+		if (_.isUndefined(propertyId)) return;
+		await appContext.deleteProperty(propertyId);
 	}, [appContext]);
 
 	return(
@@ -30,7 +28,7 @@ function AllProperties() {
 				data={appContext.Properties}
 				rightOpenValue={-50}
 				renderHiddenItem={({ item, index }) => {
-					if (appContext.Properties[index].isTenant) return null;
+					if (appContext.Properties[index].isCurrentUserTenant) return null;
 					return (
 						<TrashButton onPress={() => handleDeleteProperty(item.PropertyId)} />
 					);
